+++ RiskOfChaos/EffectDefinitions/World/Spawn/SpawnLightningStrikes.cs
@@ -1,7 +1,9 @@
 ﻿using RiskOfChaos.EffectHandling.EffectClassAttributes;
 using RiskOfChaos.Utilities;
 using RoR2;
+using RoR2.ConVar;
 using RoR2.Orbs;
+using System.Linq;
 using UnityEngine;
 using UnityEngine.AddressableAssets;
 
@@ -16,19 +18,34 @@ namespace RiskOfChaos.EffectDefinitions.World.Spawn
 
             static readonly GameObject _strikeEffect = Addressables.LoadAssetAsync<GameObject>("RoR2/Base/Lightning/LightningStrikeImpact.prefab").WaitForCompletion();
 
+            static readonly BoolConVar _enableDebugRangeIndicator = new BoolConVar("roc_lightning_strikes_debug_range_indicator", ConVarFlags.None, "0", "Enables debug visualization of the \"Risk of Thunder\" lightning strikes damage radius");
+
+            GameObject _debugRangeIndicator;
+
             public override void Begin()
             {
                 base.Begin();
 
                 duration = 1f;
+                scale = 6f;
+
                 if (_orbEffect)
                 {
                     EffectManager.SpawnEffect(_orbEffect, new EffectData
                     {
-                        scale = scale,
+                        scale = scale / 3f,
                         origin = origin,
                         genericFloat = duration
                     }, true);
+                }
+
+                if (_enableDebugRangeIndicator.value)
+                {
+                    _debugRangeIndicator = GameObject.CreatePrimitive(PrimitiveType.Sphere);
+                    _debugRangeIndicator.GetComponent<Renderer>().material.color = Color.red;
+                    _debugRangeIndicator.GetComponent<Collider>().enabled = false;
+                    _debugRangeIndicator.transform.position = origin;
+                    _debugRangeIndicator.transform.localScale = Vector3.one * scale;
                 }
             }
 
@@ -36,27 +53,33 @@ namespace RiskOfChaos.EffectDefinitions.World.Spawn
             {
                 base.OnArrival();
 
-                const float RADIUS = 3f;
-
-                EffectManager.SpawnEffect(_strikeEffect, new EffectData
+                if (_strikeEffect)
                 {
-                    origin = origin,
-                    scale = RADIUS / 3f,
-                }, true);
+                    EffectManager.SpawnEffect(_strikeEffect, new EffectData
+                    {
+                        origin = origin,
+                        scale = scale / 3f,
+                    }, true);
+                }
 
                 new BlastAttack
                 {
-                    baseDamage = 100f,
+                    baseDamage = 50f + (TeamManager.instance ? ((TeamManager.instance.GetTeamLevel(TeamIndex.Monster) - 1) * 25f) : 0f),
                     baseForce = 0f,
                     bonusForce = new Vector3(0f, 10f, 0f),
                     damageColorIndex = DamageColorIndex.Item,
-                    damageType = DamageType.SlowOnHit | DamageType.Stun1s | DamageType.Shock5s,
+                    damageType = DamageType.SlowOnHit,
                     falloffModel = BlastAttack.FalloffModel.None,
                     position = origin,
                     procCoefficient = 0f,
-                    radius = RADIUS,
+                    radius = scale,
                     teamIndex = TeamIndex.None
                 }.Fire();
+
+                if (_debugRangeIndicator)
+                {
+                    GameObject.Destroy(_debugRangeIndicator);
+                }
             }
         }
 
@@ -74,9 +97,24 @@ namespace RiskOfChaos.EffectDefinitions.World.Spawn
             {
                 if (OrbManager.instance)
                 {
+                    CharacterBody[] validStrikeTargets = CharacterBody.readOnlyInstancesList.Where(c =>
+                    {
+                        return c.healthComponent && c.healthComponent.alive && (c.bodyFlags & CharacterBody.BodyFlags.Masterless) == 0;
+                    }).ToArray();
+
+                    Vector3 origin;
+                    if (validStrikeTargets.Length > 0 && RNG.nextNormalizedFloat < Mathf.Min(0.1f, 0.0025f * validStrikeTargets.Length))
+                    {
+                        origin = RNG.NextElementUniform(validStrikeTargets).footPosition;
+                    }
+                    else
+                    {
+                        origin = SpawnUtils.EvaluateToPosition(SpawnUtils.GetBestValidRandomPlacementRule(), RNG);
+                    }
+
                     OrbManager.instance.AddOrb(new StationaryLightningStrikeOrb
                     {
-                        origin = SpawnUtils.EvaluateToPosition(SpawnUtils.GetBestValidRandomPlacementRule(), RNG),
+                        origin = origin
                     });
                 }