+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 **?.?.? Changes:**
 
+* Risk of Thunder:
+  * Increased lightning strike radius (3->6)
+  * Now properly targets characters instead of just random map locations
+  * Damage is now scaled (50 base, +25 per monster level)
+
 * Config:
   * Added mod icon to Risk of Options menu